requirements.txt
# Agentic Framework & LLM
langgraph==0.2.14
langchain==0.2.12
langchain-google-genai==1.0.8
langchain-community==0.2.11
google-generativeai==0.7.2

# Vector Database & RAG
chromadb==0.5.5
beautifulsoup4==4.12.3
requests==2.32.3
tavily-python==0.3.5

# Backend & API
fastapi==0.112.0
uvicorn==0.30.5
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0

# Utilities
rich==13.7.1
tiktoken==0.7.0
typing-extensions==4.12.2
